+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/Structure1D.kt
@@ -5,6 +5,7 @@
 
 package space.kscience.kmath.nd
 
+import space.kscience.kmath.misc.PerformancePitfall
 import space.kscience.kmath.structures.Buffer
 import space.kscience.kmath.structures.MutableBuffer
 import space.kscience.kmath.structures.asMutableBuffer
@@ -46,6 +47,8 @@ private value class Structure1DWrapper<T>(val structure: StructureND<T>) : Struc
     override val size: Int get() = structure.shape[0]
 
     override operator fun get(index: Int): T = structure[index]
+
+    @PerformancePitfall
     override fun elements(): Sequence<Pair<IntArray, T>> = structure.elements()
 }
 
@@ -55,6 +58,8 @@ private value class Structure1DWrapper<T>(val structure: StructureND<T>) : Struc
 private class MutableStructure1DWrapper<T>(val structure: MutableStructureND<T>) : MutableStructure1D<T> {
     override val shape: IntArray get() = structure.shape
     override val size: Int get() = structure.shape[0]
+
+    @PerformancePitfall
     override fun elements(): Sequence<Pair<IntArray, T>> = structure.elements()
 
     override fun get(index: Int): T = structure[index]
@@ -62,8 +67,8 @@ private class MutableStructure1DWrapper<T>(val structure: MutableStructureND<T>)
         structure[intArrayOf(index)] = value
     }
 
-    override fun copy(): MutableBuffer<T> =
-        structure.elements().map { it.second }.toMutableList().asMutableBuffer()
+    @PerformancePitfall
+    override fun copy(): MutableBuffer<T> = structure.elements().map { it.second }.toMutableList().asMutableBuffer()
 }
 
 
@@ -75,8 +80,10 @@ private value class Buffer1DWrapper<T>(val buffer: Buffer<T>) : Structure1D<T> {
     override val shape: IntArray get() = intArrayOf(buffer.size)
     override val size: Int get() = buffer.size
 
-    override fun elements(): Sequence<Pair<IntArray, T>> =
-        buffer.asSequence().mapIndexed { index, value -> intArrayOf(index) to value }
+    @PerformancePitfall
+    override fun elements(): Sequence<Pair<IntArray, T>> = buffer.asSequence().mapIndexed { index, value ->
+        intArrayOf(index) to value
+    }
 
     override operator fun get(index: Int): T = buffer[index]
 }
@@ -85,8 +92,10 @@ internal class MutableBuffer1DWrapper<T>(val buffer: MutableBuffer<T>) : Mutable
     override val shape: IntArray get() = intArrayOf(buffer.size)
     override val size: Int get() = buffer.size
 
-    override fun elements(): Sequence<Pair<IntArray, T>> =
-        buffer.asSequence().mapIndexed { index, value -> intArrayOf(index) to value }
+    @PerformancePitfall
+    override fun elements(): Sequence<Pair<IntArray, T>> = buffer.asSequence().mapIndexed { index, value ->
+        intArrayOf(index) to value
+    }
 
     override operator fun get(index: Int): T = buffer[index]
     override fun set(index: Int, value: T) {