+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/Structure2D.kt
@@ -5,10 +5,11 @@
 
 package space.kscience.kmath.nd
 
+import space.kscience.kmath.misc.PerformancePitfall
 import space.kscience.kmath.misc.UnstableKMathAPI
 import space.kscience.kmath.structures.Buffer
-import space.kscience.kmath.structures.VirtualBuffer
 import space.kscience.kmath.structures.MutableListBuffer
+import space.kscience.kmath.structures.VirtualBuffer
 import kotlin.jvm.JvmInline
 import kotlin.reflect.KClass
 
@@ -33,12 +34,14 @@ public interface Structure2D<T> : StructureND<T> {
     /**
      * The buffer of rows of this structure. It gets elements from the structure dynamically.
      */
+    @PerformancePitfall
     public val rows: List<Buffer<T>>
         get() = List(rowNum) { i -> VirtualBuffer(colNum) { j -> get(i, j) } }
 
     /**
      * The buffer of columns of this structure. It gets elements from the structure dynamically.
      */
+    @PerformancePitfall
     public val columns: List<Buffer<T>>
         get() = List(colNum) { j -> VirtualBuffer(rowNum) { i -> get(i, j) } }
 
@@ -80,12 +83,14 @@ public interface MutableStructure2D<T> : Structure2D<T>, MutableStructureND<T> {
     /**
      * The buffer of rows of this structure. It gets elements from the structure dynamically.
      */
+    @PerformancePitfall
     override val rows: List<MutableStructure1D<T>>
         get() = List(rowNum) { i -> MutableBuffer1DWrapper(MutableListBuffer(colNum) { j -> get(i, j) })}
 
     /**
      * The buffer of columns of this structure. It gets elements from the structure dynamically.
      */
+    @PerformancePitfall
     override val columns: List<MutableStructure1D<T>>
         get() = List(colNum) { j -> MutableBuffer1DWrapper(MutableListBuffer(rowNum) { i -> get(i, j) }) }
 }
@@ -105,6 +110,7 @@ private value class Structure2DWrapper<T>(val structure: StructureND<T>) : Struc
     @UnstableKMathAPI
     override fun <F : StructureFeature> getFeature(type: KClass<out F>): F? = structure.getFeature(type)
 
+    @PerformancePitfall
     override fun elements(): Sequence<Pair<IntArray, T>> = structure.elements()
 }